fix deleting location point

+++ ItPlanet.Web/Services/LocationPoint/LocationPointService.cs
@@ -38,6 +38,8 @@ public class LocationPointService : ILocationPointService
     public async Task DeletePointAsync(long pointId)
     {
         var point = await GetLocationPointAsync(pointId);
+        if (await _repository.HasLinkedAnimal(pointId) || point.VisitedPoints.Any())
+            throw new UnableDeleteLocationPointException();
         await _repository.DeleteAsync(point).ConfigureAwait(false);
     }
